+++ backend/app/api/routes/login.py
@@ -10,7 +10,7 @@ from app.core import security
 from app.core.config import settings
 from app.core.security import get_password_hash
 from app.model.users import Message, NewPassword, Token, UserOut
-from app.repository import users
+from app.crud import users
 from app.utils import (
     generate_password_reset_token,
     generate_reset_password_email,
