+++ backend/app/tests/api/routes/test_users.py
@@ -5,7 +5,7 @@ from sqlmodel import Session
 
 from app.core.config import settings
 from app.model.users import UserCreate
-from app.repository import users
+from app.crud import users
 from app.tests.utils.utils import random_email, random_lower_string
