+++ backend/app/tests/utils/item.py
@@ -1,7 +1,7 @@
 from sqlmodel import Session
 
 from app.model.users import Item, ItemCreate
-from app.repository import users
+from app.crud import users
 from app.tests.utils.user import create_random_user
 from app.tests.utils.utils import random_lower_string
